requirements.txt
pandas
numpy
scipy
scikit-learn
lightgbm
joblib
pyyaml
optuna
matplotlib
